@@ -7,7 +7,7 @@
     "start": "webpack-dev-server",
     "build": "./node_modules/webpack/bin/webpack.js",
     "test": "jest",
-    "eslint": "./node_modules/eslint/bin/eslint.js ./lib/*.js"
+    "eslint": "./node_modules/eslint/bin/eslint.js ./lib/"
   },
   "repository": {
     "type": "git",
@@ -31,7 +31,7 @@
   "devDependencies": {
     "babel": "^6.5.2",
     "babel-core": "^6.24.1",
-    "babel-eslint": "^6.0.4",
+    "babel-eslint": "^8.1.0",
     "babel-jest": "^20.0.3",
     "babel-loader": "^6.2.4",
     "babel-polyfill": "^6.7.4",
@@ -41,7 +41,7 @@
     "chai": "^3.5.0",
     "css-loader": "^0.23.1",
     "enzyme": "^2.8.2",
-    "eslint": "^2.6.0",
+    "eslint": "^4.18.1",
     "eslint-config-airbnb": "^9.0.1",
     "eslint-plugin-import": "^1.8.0",
     "eslint-plugin-jsx-a11y": "^1.2.2",
