@@ -5,7 +5,6 @@ import SevenHour from './SevenHour';
 import TenDay from './TenDay';
 import Welcome from './Welcome';
 import ErrorPage from './Error';
-import data from '../data/Mock-data';
 import retrieveWeather from '../data/api-data';
 import getCurrentWeather from '../data-cleaners/current-weather-getter';
 import getTenDay from '../data-cleaners/ten-day-getter';
@@ -13,25 +12,18 @@ import getSevenHour from '../data-cleaners/seven-hour-getter';
 import getBackgroundState from '../data-cleaners/background-images-selector';
 import '../../styles/App.scss';
 
-
-const currentWeather = getCurrentWeather(data);
-const tenDay = getTenDay(data);
-const sevenHour = getSevenHour(data);
-const backgroundImage = getBackgroundState(data);
-
-
 class App extends React.Component {
   constructor() {
-    super()
+    super();
     this.state = {
       data: null,    
-      currentWeather: '',
-      sevenHour: '',
-      tenDay: '',
+      currentWeather: {},
+      sevenHour: [],
+      tenDay: [],
       location: '',
       background: '',
       error: false
-    }
+    };
     this.updateWeather = this.updateWeather.bind(this);
     this.getFromStorage = this.getFromStorage.bind(this);
     this.sendLocationToStorage = this.sendLocationToStorage.bind(this);
@@ -43,26 +35,30 @@ class App extends React.Component {
     }
   }
 
-  updateWeather(location) {
-    retrieveWeather(location)
+  updateWeather(locationInput) {
+    retrieveWeather(locationInput)
       .then(response => response.json())
       .then(weatherInfo => {
-        if(weatherInfo.forecast) {
-          this.sendLocationToStorage(location)
+        if (weatherInfo.forecast) {
+          this.sendLocationToStorage(locationInput);
           this.setState({
             data: weatherInfo, 
             currentWeather: getCurrentWeather(weatherInfo), 
             sevenHour: getSevenHour(weatherInfo), 
             tenDay: getTenDay(weatherInfo),
-            location: location,
+            location: locationInput,
             background: getBackgroundState(weatherInfo),
             error: false
-          })
+          });
         } else {
-          this.setState({error: true})
+          this.setState({error: true});
         }
       })
-      .catch(error => {this.setState( {error: true} ) })  
+      .catch(error => {
+        this.setState( {
+          error: true
+        } ); 
+      }); 
   }
 
   sendLocationToStorage(location) {
@@ -71,14 +67,15 @@ class App extends React.Component {
 
   getFromStorage() {
     let retrievedData = localStorage.getItem('location');
+
     this.updateWeather(retrievedData);
   }
 
   render () {
-    if(this.state.error === false) {
-    return (
-      <div>
-        {this.state.data ? 
+    if (this.state.error === false) {
+      return (
+        <div>
+          {this.state.data ? 
             (<main className={this.state.background}>
               <section className="main-banner">
                 <Header updateWeather={this.updateWeather}/>
@@ -89,17 +86,17 @@ class App extends React.Component {
               </section>
               <TenDay array={this.state.tenDay}/>
             </main>) :     
-          (<Welcome updateWeather={this.updateWeather}/>)
-      }
-      </div>
-    )
-  }else {
-    return (
-      <div>
-        <ErrorPage updateWeather={this.updateWeather}/>
-      </div>
-    )
-  }
+            (<Welcome updateWeather={this.updateWeather}/>)
+          }
+        </div>
+      );
+    } else {
+      return (
+        <div>
+          <ErrorPage updateWeather={this.updateWeather}/>
+        </div>
+      );
+    }
   }
 }
 
