@@ -1,5 +1,6 @@
 import React from 'react';
 import CardContent from './CardContent';
+import PropTypes from 'prop-types';
 import '../../styles/Card.scss';
 
 
@@ -7,14 +8,19 @@ const Card = (props) => {
   return (
     <article className={props.cardClass}>
       <ul>
-      {
-        props.array.map((item, index) => {
-          return <CardContent key={index} weatherInfo={item}/>
-        })
-      }
+        {
+          props.array.map((weatherInfo, index) => {
+            return <CardContent key={index} weatherInfo={weatherInfo}/>;
+          })
+        }
       </ul>
     </article>
-  )
-}
+  );
+};
+
+Card.propTypes = {
+  cardClass: PropTypes.string,
+  array: PropTypes.array
+};
 
 export default Card;
