@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import '../../styles/CurrentWeather.scss';
 
 
@@ -12,15 +13,24 @@ const CurrentWeather = (props) =>  {
         </div>
         <div className="local-weather-div-two">
           <h5>{props.forecast.currentTime[0]}</h5>
-          <h5>{props.forecast.currentTime[1] + ' ' + props.forecast.currentTime[2]}</h5>
-          <h5>{props.forecast.currentCondition}<img src={props.forecast.icon}/></h5>
-          <p className="temp">{props.forecast.highTemp + '°'}/<span className="low">{props.forecast.lowTemp + '°'}</span></p> 
+          <h5>{props.forecast.currentTime[1] +
+           ' ' + props.forecast.currentTime[2]}</h5>
+          <h5>{props.forecast.currentCondition}
+            <img src={props.forecast.icon}/>  
+          </h5>
+          <p className="temp">{props.forecast.highTemp + '°'}/
+          <span className="low">{props.forecast.lowTemp + '°'}</span>
+          </p> 
           <p className="forecast">{props.forecast.summary}</p>
         </div>  
       </div>
     </article>
-  )
-}
+  );
+};
+
+CurrentWeather.propTypes = {
+  forecast: PropTypes.object
+};
 
 
 
