@@ -1,6 +1,7 @@
 import React from 'react';
 import Search from './Search';
-import '../../styles/Error.scss'
+import PropTypes from 'prop-types';
+import '../../styles/Error.scss';
 
 const ErrorPage = (props) => {
   return (
@@ -11,7 +12,11 @@ const ErrorPage = (props) => {
         <Search updateWeather={props.updateWeather}/>
       </div>
     </div>
-  )
-}
+  );
+};
+
+ErrorPage.propTypes = {
+  updateWeather: PropTypes.func
+};
 
 export default ErrorPage;
