@@ -1,6 +1,6 @@
 import React from 'react';
 import Search from './Search';
-import MobileMenu from './MobileSearch'
+import PropTypes from 'prop-types';
 import '../../styles/Header.scss';
 
 
@@ -9,9 +9,12 @@ const Header = (props) => {
     <header>
       <h2>Weathrly</h2>
       <Search updateWeather={props.updateWeather}/>
-      <MobileMenu />
     </header>
-  )
-}
+  );
+};
+
+Header.propTypes = {
+  updateWeather: PropTypes.func
+};
 
 export default Header;
