@@ -1,5 +1,5 @@
 import React from 'react';
-import '../../styles/MobileMenu.scss'
+import '../../styles/MobileMenu.scss';
 
 
 const MobileMenu = () => {
@@ -9,7 +9,7 @@ const MobileMenu = () => {
       <li>Hourly Forecast</li>
       <li>Daily Forecast</li>
     </ul>
-  )
-}
+  );
+};
 
 export default MobileMenu;
