@@ -1,14 +1,15 @@
 import React from 'react';
-import Search from './Search'
+import Search from './Search';
+import PropTypes from 'prop-types';
 import '../../styles/Welcom.scss';
 
 
 class Welcome extends React.Component  {
   constructor(props) {
-    super(props)
+    super(props);
     this.state = {
       time: new Date().getHours()
-    }
+    };
   }
   render () {
     return (
@@ -19,7 +20,8 @@ class Welcome extends React.Component  {
             this.state.time < 12 && <p className="greeting">Good Morning</p>
           }
           {
-            this.state.time >= 12 && this.state.time < 18 &&<p className="greeting">Good Afternoon</p>
+            this.state.time >= 12 && this.state.time < 18 
+            && <p className="greeting">Good Afternoon</p>
           }
           {
             this.state.time >= 18 && <p className="greeting">Good Evening</p>
@@ -27,8 +29,12 @@ class Welcome extends React.Component  {
           <Search updateWeather={this.props.updateWeather}/>
         </div>
       </div>
-    )
+    );
   }
 }
+
+Welcome.propTypes = {
+  updateWeather: PropTypes.func
+};
 
 export default Welcome;
