@@ -1,16 +1,19 @@
 const getTenDay = (data) => {
-  const tenDay = data.forecast.simpleforecast.forecastday.reduce((array, day) => {
-  array.push(
+  let tenDay = data.forecast.simpleforecast.forecastday;
+  let tenDayForecast;
+
+  tenDayForecast = tenDay.reduce((tenDayArray, day) => {
+    tenDayArray.push(
       {
         day: day.date.weekday,
         img: day.icon_url,
         high: `${day.high.fahrenheit}°`,
         low: `${day.low.fahrenheit}°`
       }
-    )
-    return array;
+    );
+    return tenDayArray;
   }, []);
-  return tenDay;
-}
+  return tenDayForecast;
+};
 
 export default getTenDay;
