@@ -49,9 +49,9 @@ const data = {
     local_tz_short: 'EST',
     local_tz_long: 'America/New_York',
     local_tz_offset: '-0500',
-    weather: 'Thunderstorms with Small Hail',
+    weather: 'Sand',
     temperature_string: '46.0 F (7.8 C)',
-    temp_f: 46.8,
+    temp_f: 30,
     temp_c: 7.8,
     relative_humidity: '65%',
     wind_string: 'From the NE at 1.0 MPH Gusting to 8.0 MPH',
@@ -324,7 +324,7 @@ const data = {
             month: 12,
             year: 2017,
             yday: 353,
-            hour: 19,
+            hour: 12,
             min: '00',
             sec: 0,
             isdst: '0',
@@ -2862,4 +2862,4 @@ const data = {
     },
   ],
 };
- export default data;
+export default data;
