@@ -1,10 +1,10 @@
 import key from '../key';
 
-const proxyurl = "https://cors-anywhere.herokuapp.com/";
 const root = 'http://api.wunderground.com';
+const conditions = 'conditions/forecast10day/hourly/q';
 
 const retrieveWeather = (location) => {
-  return fetch(`${root}/api/${key}/conditions/forecast10day/hourly/q/${location}.json`)
-}
+  return fetch(`${root}/api/${key}/${conditions}/${location}.json`);
+};
 
 export default retrieveWeather;
