@@ -1001,7 +1001,7 @@ const cityList = {
     "Beloit, WI",
     "Panama City, FL"
   ]
-}
+};
 
 
 export default cityList;
