@@ -1,6 +1,11 @@
 import React from 'react';
 import { shallow, mount } from 'enzyme';
 import App from '../lib/components/App';
+import PropTypes from 'prop-types';
+import data from '../lib/data/Mock-data';
+import getCurrentWeather from '../lib/data-cleaners/current-weather-getter';
+import getSevenHour from '../lib/data-cleaners/seven-hour-getter';
+import getTenDay from '../lib/data-cleaners/ten-day-getter';
 
 describe('App component shallow', () => {
   let wrapper;
@@ -17,15 +22,15 @@ describe('App component shallow', () => {
   })
 
   it('should initially start with a state of currentWeather set to an empty string', () => {
-    expect(wrapper.state('currentWeather')).toEqual( '');
+    expect(wrapper.state('currentWeather')).toEqual( {});
   })
 
   it('should initially start with a state of sevenHour set to an empty string', () => {
-    expect(wrapper.state('sevenHour')).toEqual('');
+    expect(wrapper.state('sevenHour')).toEqual([]);
   })
 
   it('should initially start with a state of tenDay set to an empty string', () => {
-    expect(wrapper.state('tenDay')).toEqual( '' );
+    expect(wrapper.state('tenDay')).toEqual( [] );
   })
 
   it('should initially start with a state of location set to an empty string', () => {
@@ -45,22 +50,22 @@ describe('App component shallow', () => {
   })
 
   it('should render the Header component if location data is present', () => {
-    wrapper.instance().setState({data: true});
+    wrapper.instance().setState({data});
     expect(wrapper.find('Header').length).toEqual(1);
   })
 
   it('should render the CurrentWeather component if location data is present', () => {
-    wrapper.instance().setState({data: true});
+    wrapper.instance().setState({data});
     expect(wrapper.find('CurrentWeather').length).toEqual(1);
   })
   
   it('should render the SevenHour component if location data is present', () => {
-    wrapper.instance().setState({data: true});
+    wrapper.instance().setState({data});
     expect(wrapper.find('SevenHour').length).toEqual(1);
   })
 
   it('should render the TenDay component if location data is present', () => {
-    wrapper.instance().setState({data: true});
+    wrapper.instance().setState({data});
     expect(wrapper.find('TenDay').length).toEqual(1);
   })
 
